+++ Ceen.Database/IDatabaseDialect.cs
@@ -99,14 +99,13 @@ namespace Ceen.Database
         /// <param name="offset">The optional offset to use</param>
         /// <param name="limit">The maximum number of items to use</param>
         /// <returns>The limit fragment</returns>
-        string Limit(int limit, int? offset = null);
+        string Limit(long limit, long? offset = null);
 
         /// <summary>
         /// Renders an SQL where clause from a query element
         /// </summary>
         /// <param name="type">The type to generate the clause for.</param>
         /// <param name="element">The element to use</param>
-        /// <param name="order">The query order</param>
         /// <returns>The SQL where clause</returns>
         KeyValuePair<string, object[]> RenderWhereClause(Type type, QueryElement element);
